refactor: update chart common name
