+++ src/pages/StatisticsChart/common/index.tsx
