+++ src/pages/StatisticsChart/nervosDao/NewDaoDeposit.tsx
@@ -10,7 +10,7 @@ import { isMobile } from '../../../utils/screen'
 import { shannonToCkb } from '../../../utils/util'
 import { ChartColors } from '../../../utils/const'
 import { isMainnet } from '../../../utils/chain'
-import { ChartLoading, ReactChartCore, ChartPage, tooltipWidth, tooltipColor } from '../common/ChartComp'
+import { ChartLoading, ReactChartCore, ChartPage, tooltipWidth, tooltipColor } from '../common'
 import { AppDispatch } from '../../../contexts/reducer'
 import { PageActions } from '../../../contexts/actions'
